backend/requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.0.3
SQLAlchemy==2.0.22
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0
gunicorn==20.1.0
